Made a parser for my own use
This is the first public source release for this project.
It's mainly a lazy attempt at making Markdown files a feasable way to
submit things in PDFs (Markdown -> Web Browser -> PDF) as opposed to using
LaTeX or a fully fledged word process (Word or Docs, pick your poison).

I'm fond of Markdown files so I'll use that.